+++ .docgenirc.js
@@ -3,7 +3,7 @@ module.exports = {
     title: 'TETHYS',
     siteProjectName: 'site',
     repoUrl: 'https://github.com/atinc/ngx-tethys',
-    logoUrl: 'https://cdn.worktile.com/open-sources/ngx-tethys/logos/tethys.png',
+    logoUrl: 'https://cdn.worktile.com/open-sources/ngx-tethys/logos/tethys.png?100',
     navs: [
         null,
         {
@@ -102,9 +102,16 @@ module.exports = {
                         }
                     }
                 }
-            ]
+            ],
+            labels: {
+                'lack-doc': {
+                    text: 'Lack Doc',
+                    color: '#2dbcff'
+                }
+            }
         }
     ],
+    footer: "Copyright © 2020-present Powered by <a href='https://worktile.com' target='_blank' >Worktile</a>",
     locales: [
         {
             key: 'zh-cn',