+++ src/date-picker/range-picker.component.spec.ts
@@ -10,7 +10,7 @@ import { FormsModule } from '@angular/forms';
 import { By } from '@angular/platform-browser';
 
 import { ThyDatePickerModule } from './date-picker.module';
-import { RangeEntry } from './standard-types';
+import { RangeEntry, PanelMode } from './standard-types';
 
 registerLocaleData(zh);
 
@@ -284,29 +284,54 @@ describe('ThyRangePickerComponent', () => {
         beforeEach(() => (fixtureInstance.useSuite = 1));
 
         it('should support date panel changes', fakeAsync(() => {
-            fixtureInstance.modelValue = { begin: new Date('2018-6-11'), end: new Date('2018-12-12') };
+            fixtureInstance.modelValue = { begin: new Date('2018-6-11'), end: new Date('2020-12-12') };
             fixture.detectChanges();
             openPickerByClickTrigger();
             // Click previous year button
             dispatchMouseEvent(queryFromOverlay('.thy-calendar-range-left .thy-calendar-prev-year-btn'), 'click');
             fixture.detectChanges();
-            expect(queryFromOverlay('.thy-calendar-range-left .thy-calendar-year-select').textContent.indexOf('2017') > -1).toBeTruthy();
+            expect(queryFromOverlay('.thy-calendar-range-left .thy-calendar-year-btn').textContent.indexOf('2017') > -1).toBeTruthy();
             // Click next year button * 2
             dispatchMouseEvent(queryFromOverlay('.thy-calendar-range-left .thy-calendar-next-year-btn'), 'click');
             fixture.detectChanges();
             dispatchMouseEvent(queryFromOverlay('.thy-calendar-range-left .thy-calendar-next-year-btn'), 'click');
             fixture.detectChanges();
-            expect(queryFromOverlay('.thy-calendar-range-left .thy-calendar-year-select').textContent.indexOf('2019') > -1).toBeTruthy();
+            expect(queryFromOverlay('.thy-calendar-range-left .thy-calendar-year-btn').textContent.indexOf('2019') > -1).toBeTruthy();
             // Click previous month button
             dispatchMouseEvent(queryFromOverlay('.thy-calendar-range-left .thy-calendar-prev-month-btn'), 'click');
             fixture.detectChanges();
-            expect(queryFromOverlay('.thy-calendar-range-left .thy-calendar-month-select').textContent.indexOf('5') > -1).toBeTruthy();
+            expect(queryFromOverlay('.thy-calendar-range-left .thy-calendar-month-btn').textContent.indexOf('5') > -1).toBeTruthy();
             // Click next month button * 2
             dispatchMouseEvent(queryFromOverlay('.thy-calendar-range-left .thy-calendar-next-month-btn'), 'click');
             fixture.detectChanges();
             dispatchMouseEvent(queryFromOverlay('.thy-calendar-range-left .thy-calendar-next-month-btn'), 'click');
             fixture.detectChanges();
-            expect(queryFromOverlay('.thy-calendar-range-left .thy-calendar-month-select').textContent.indexOf('7') > -1).toBeTruthy();
+            expect(queryFromOverlay('.thy-calendar-range-left .thy-calendar-month-btn').textContent.indexOf('7') > -1).toBeTruthy();
+        }));
+
+        it('should show current thy-calendar-next-month-btn and thy-calendar-next-year-btn', fakeAsync(() => {
+            fixtureInstance.modelValue = { begin: new Date('2018-10-11'), end: new Date('2018-12-12') };
+            fixture.detectChanges();
+            openPickerByClickTrigger();
+            expect(queryFromOverlay('.thy-calendar-range-left .thy-calendar-next-year-btn')).toBeFalsy();
+            expect(queryFromOverlay('.thy-calendar-range-right .thy-calendar-prev-year-btn')).toBeFalsy();
+            fixture.detectChanges();
+            dispatchMouseEvent(queryFromOverlay('.thy-calendar-range-left  .thy-calendar-header .thy-calendar-next-month-btn'), 'click');
+            fixture.detectChanges();
+            tick(500);
+            fixture.detectChanges();
+            fixture.detectChanges();
+            expect(queryFromOverlay('.thy-calendar-range-left .thy-calendar-next-month-btn')).toBeFalsy();
+            expect(queryFromOverlay('.thy-calendar-range-right .thy-calendar-prev-month-btn')).toBeFalsy();
+            fixture.detectChanges();
+            dispatchMouseEvent(queryFromOverlay('.thy-calendar-range-left .thy-calendar-prev-year-btn'), 'click');
+            fixture.detectChanges();
+            tick(500);
+            fixture.detectChanges();
+            expect(queryFromOverlay('.thy-calendar-range-left .thy-calendar-next-month-btn')).toBeDefined();
+            expect(queryFromOverlay('.thy-calendar-range-right .thy-calendar-prev-month-btn')).toBeDefined();
+            expect(queryFromOverlay('.thy-calendar-range-left .thy-calendar-next-year-btn')).toBeDefined();
+            expect(queryFromOverlay('.thy-calendar-range-right .thy-calendar-pre-year-btn')).toBeDefined();
         }));
     }); // /panel switch and move forward/afterward
 
@@ -351,6 +376,56 @@ describe('ThyRangePickerComponent', () => {
         }));
     });
 
+    describe('month mode', () => {
+        beforeEach(() => {
+            fixtureInstance.useSuite = 1;
+            fixtureInstance.thyMode = 'month';
+        });
+
+        it('should show month panel', fakeAsync(() => {
+            fixtureInstance.modelValue = { begin: new Date('2018-09-11'), end: new Date('2020-10-12') };
+            fixture.detectChanges();
+            openPickerByClickTrigger();
+            expect(queryFromOverlay('.thy-calendar-range-left .thy-calendar-month')).toBeTruthy();
+            expect(queryFromOverlay('.thy-calendar-range-left .thy-calendar-month-btn').textContent.trim()).toEqual('2018年');
+            expect(queryFromOverlay('.thy-calendar-range-right .thy-calendar-month')).toBeTruthy();
+            expect(queryFromOverlay('.thy-calendar-range-right .thy-calendar-month-btn').textContent.trim()).toEqual('2020年');
+        }));
+        it('should support previous and next year', fakeAsync(() => {
+            fixtureInstance.modelValue = { begin: new Date('2018-09-11'), end: new Date('2018-10-12') };
+            fixture.detectChanges();
+            openPickerByClickTrigger();
+            fixture.detectChanges();
+            tick(500);
+            fixture.detectChanges();
+            expect(queryFromOverlay('.thy-calendar-range-left .thy-calendar-month-btn').textContent.indexOf('2018') > -1).toBeTruthy();
+            expect(queryFromOverlay('.thy-calendar-range-right .thy-calendar-month-btn').textContent.indexOf('2019') > -1).toBeTruthy();
+            expect(queryFromOverlay('.thy-calendar-range-left .thy-calendar-next-year-btn')).not.toBeTruthy();
+            expect(queryFromOverlay('.thy-calendar-range-right .thy-calendar-prev-year-btn')).not.toBeTruthy();
+            dispatchMouseEvent(queryFromOverlay('.thy-calendar-range-left .thy-calendar-prev-year-btn'), 'click');
+            fixture.detectChanges();
+            expect(queryFromOverlay('.thy-calendar-range-left .thy-calendar-month-btn').textContent.indexOf('2017') > -1).toBeTruthy();
+            fixture.detectChanges();
+            expect(queryFromOverlay('.thy-calendar-range-left .thy-calendar-next-year-btn')).toBeTruthy();
+            expect(queryFromOverlay('.thy-calendar-range-right .thy-calendar-prev-year-btn')).toBeTruthy();
+            fixture.detectChanges();
+            tick(500);
+            dispatchMouseEvent(queryFromOverlay('.thy-calendar-range-left .thy-calendar-next-year-btn'), 'click');
+            fixture.detectChanges();
+            expect(queryFromOverlay('.thy-calendar-range-left .thy-calendar-month-btn').textContent.indexOf('2018') > -1).toBeTruthy();
+        }));
+        it('should support panel changes', fakeAsync(() => {
+            fixtureInstance.modelValue = { begin: new Date('2018-09-11'), end: new Date('2018-10-12') };
+            fixture.detectChanges();
+            openPickerByClickTrigger();
+            fixture.detectChanges();
+            tick(500);
+            fixture.detectChanges();
+            dispatchMouseEvent(queryFromOverlay('.thy-calendar-range-left .thy-calendar-month-btn'), 'click');
+            fixture.detectChanges();
+            expect(queryFromOverlay('.thy-calendar-range-left .thy-calendar-year')).toBeTruthy();
+        }));
+    });
     ////////////
 
     function getPickerTrigger(): HTMLInputElement {
@@ -404,6 +479,7 @@ describe('ThyRangePickerComponent', () => {
                 [thySuffixIcon]="thySuffixIcon"
                 (thyOpenChange)="thyOpenChange($event)"
                 [(ngModel)]="modelValue"
+                [thyMode]="thyMode"
                 (ngModelChange)="modelValueChange($event)"
                 (thyOnPanelChange)="thyOnPanelChange($event)"
                 (thyOnCalendarChange)="thyOnCalendarChange($event)"
@@ -434,6 +510,7 @@ class ThyTestRangePickerComponent {
     thySize: string;
     thySuffixIcon: string;
     modelValue: RangeEntry;
+    thyMode: PanelMode;
     thyOpen: boolean;
     thyOpenChange(): void {}
     modelValueChange(): void {}