+++ src/dialog/api/zh-cn.js
@@ -1,12 +1,80 @@
 module.exports = [
     {
-        type: 'CLASS',
-        name: 'Config参数',
+        type: 'service',
+        name: 'ThyDialog',
         properties: [
             {
+                name: 'open',
+                description: '打开 Dialog',
+                type: `(
+                    componentOrTemplateRef: ComponentType<T> | TemplateRef<T>,
+                    config?: ThyDialogConfig
+                ) => ThyDialogRef`,
+                default: ''
+            },
+            {
+                name: 'confirm',
+                description: '打开 Confirm',
+                type: `(options: ThyConfirmConfig) => ThyDialogRef`,
+                default: ''
+            },
+            {
+                name: 'getDialogById',
+                description: '根据 id 获取 Dialog',
+                type: `(id: string) => ThyDialogRef`,
+                default: ''
+            },
+            {
+                name: 'getOpenedDialogs',
+                description: '获取所有打开的 Dialog',
+                type: `() => ThyDialogRef[]`,
+                default: ''
+            },
+            {
+                name: 'getClosestDialog',
+                description: '获取与指定元素最接近的 ThyDialogRef',
+                type: `(element: HTMLElement) => ThyDialogRef`,
+                default: ''
+            },
+            {
+                name: 'afterOpened',
+                description: 'Dialog 打开后的回调',
+                type: '() => Subject',
+                default: ''
+            },
+            {
+                name: 'close',
+                description: '关闭 Dialog',
+                type: '() => void',
+                default: ''
+            },
+            {
+                name: 'closeAll',
+                description: '关闭所有打开的 Dialog',
+                type: '() => void',
+                default: ''
+            },
+            {
+                name: 'afterAllClosed',
+                description: '所有 Dialog 完全关闭后的回调',
+                type: '() => Subject',
+                default: ''
+            }
+        ]
+    },
+    {
+        type: 'CLASS',
+        name: 'ThyDialogConfig',
+        properties: [
+            {
+                name: 'id',
+                description: '弹出框的唯一标识',
+                type: 'string'
+            },
+            {
                 name: 'size',
-                description: '弹出框的大小，ThyDialogSizes: sm, md, lg, max-lg',
-                type: 'ThyDialogSizes | string',
+                description: '弹出框的大小，ThyDialogSizes: sm (400), md (660), lg (800), maxLg (980), supperLg (94vw), full (全屏)',
+                type: 'ThyDialogSizes',
                 default: 'md'
             },
             {
@@ -20,6 +88,11 @@ module.exports = [
                 description: '点击幕布或者按ESC键是否自动关闭弹出框',
                 type: 'boolean',
                 default: 'true'
+            },
+            {
+                name: 'backdropClass',
+                description: '自定义幕布的样式',
+                type: 'string | string[]'
             },
             {
                 name: 'closeOnNavigation',
@@ -44,6 +117,71 @@ module.exports = [
                 description: '弹出框最大高度',
                 type: 'number | string',
                 default: '85vh'
+            },
+            {
+                name: 'width',
+                description: '自定义弹出框的宽度',
+                type: 'string',
+                default: '660px'
+            },
+            {
+                name: 'height',
+                description: '自定义弹出框的高度',
+                type: 'string',
+                default: '85vh'
+            },
+            {
+                name: 'position',
+                description: '定位模态框的弹出位置',
+                type: '{ top?: string; bottom?: string; left?: string; right?: string;}'
+            }
+        ]
+    },
+    {
+        type: 'interface',
+        name: 'ThyDialogRef',
+        properties: [
+            {
+                name: 'close',
+                description: '关闭当前 Dialog',
+                type: '() => void',
+                default: ''
+            },
+            {
+                name: 'afterOpened',
+                description: 'Dialog 打开后的回调',
+                type: '() => Observable',
+                default: ''
+            },
+            {
+                name: 'beforeClosed',
+                description: 'Dialog 关闭前的回调',
+                type: '() => Observable',
+                default: ''
+            },
+            {
+                name: 'afterClosed',
+                description: 'Dialog 关闭后的回调',
+                type: '() => Observable',
+                default: ''
+            },
+            {
+                name: 'backdropClick',
+                description: '点击 Dialog 遮罩层的回调',
+                type: '() => Observable',
+                default: ''
+            },
+            {
+                name: 'getOverlayRef',
+                description: '获取 ThyDialogRef',
+                type: '() => ThyDialogRef',
+                default: ''
+            },
+            {
+                name: 'updatePosition',
+                description: '更新 Dialog 的位置',
+                type: '() => ThyDialogRef',
+                default: ''
             }
         ]
     },
@@ -72,7 +210,7 @@ module.exports = [
             {
                 name: '#dialogHeader',
                 description: '自定义头部模版',
-                type: 'TemplateRef',
+                type: 'ContentChild<TemplateRef>',
                 default: ''
             },
             {
@@ -101,7 +239,7 @@ module.exports = [
         properties: [
             {
                 name: 'thyDivided',
-                description: '顶部是否有分割线，可全局配置默认值',
+                description: '底部是否有分割线，可全局配置默认值',
                 type: 'boolean',
                 default: 'false'
             },
@@ -110,6 +248,66 @@ module.exports = [
                 description: '对齐方式，`left | center | right`，可全局配置默认值',
                 type: 'string',
                 default: 'left'
+            },
+            {
+                name: '#description',
+                description: '自定义弹出框底部的描述模板',
+                type: 'ContentChild<TemplateRef>',
+                default: ''
+            }
+        ]
+    },
+    {
+        type: 'CLASS',
+        name: 'ThyConfirmConfig',
+        properties: [
+            {
+                name: 'title',
+                description: '标题',
+                type: 'string',
+                default: '确认删除'
+            },
+            {
+                name: 'content',
+                description: '提示内容',
+                type: 'string',
+                default: ''
+            },
+            {
+                name: 'okText',
+                description: '确认按钮的文案',
+                type: 'string',
+                default: '确认'
+            },
+            {
+                name: 'cancelText',
+                description: '取消按钮的文案',
+                type: 'string',
+                default: '取消'
+            },
+            {
+                name: 'okType',
+                description: '确认按钮的类型，primary | danger',
+                type: 'string',
+                default: 'danger'
+            },
+            {
+                name: 'okLoadingText',
+                description: '确认按钮处于提交状态时的文案',
+                type: 'string',
+                default: 'okText'
+            },
+            {
+                name: 'footerAlign',
+                description: '底部对齐方式，left | center | right',
+                type: 'string',
+                default: 'left'
+            },
+            {
+                name: 'onOk',
+                description: '确认后的回调事件',
+                type: '() => Observable<boolean> | void',
+                default: ''
             }
         ]
     }