+++ src/dialog/examples/module.ts
@@ -8,8 +8,20 @@ import { ThyDialogBasicExampleComponent } from './basic/basic.component';
 import { ThyDialogBasicContentComponent } from './basic/dialog-content.component';
 import { ThyDialogConfirmExampleComponent } from './confirm/confirm.component';
 import { CdkScrollableModule } from '@angular/cdk/scrolling';
+import { ThyDialogSizeExampleComponent } from './size/size.component';
+import { ThyDialogBackdropExampleComponent } from './backdrop/backdrop.component';
+import { ThyDialogLayoutExampleComponent } from './layout/layout.component';
+import { ThyDialogFooterLayoutExampleComponent } from './layout/dialog-layout.component';
 
-const COMPONENTS = [ThyDialogBasicExampleComponent, ThyDialogBasicContentComponent, ThyDialogConfirmExampleComponent];
+const COMPONENTS = [
+    ThyDialogBasicExampleComponent,
+    ThyDialogBasicContentComponent,
+    ThyDialogConfirmExampleComponent,
+    ThyDialogSizeExampleComponent,
+    ThyDialogBackdropExampleComponent,
+    ThyDialogLayoutExampleComponent,
+    ThyDialogFooterLayoutExampleComponent
+];
 
 @NgModule({
     declarations: COMPONENTS,